Add a total_cmp method to Measure and MeasurePoint
This method is generic over the measurement unit, but the value
type is restricted to be either `f32` or `f64`, since these types
have a `total_cmp` method that can be deferred to.

With this method, it becomes possible to sort a slice of measures
(or measure points) using `sort_by`. For example:

```
let mut m_array = [
    Measure::<Metre, f64>::new(31.),
    Measure::<Metre, f64>::new(12.),
    Measure::<Metre, f64>::new(43.),
    Measure::<Metre, f64>::new(24.),
];
m_array.sort_by(Measure::<Metre>::total_cmp);
assert_eq!(m_array, [
    Measure::<Metre>::new(12.),
    Measure::<Metre>::new(24.),
    Measure::<Metre>::new(31.),
    Measure::<Metre>::new(43.),
]);
```